requirements.txt
pdf2image
pillow
pytesseract
faiss-cpu 
sentence-transformers
google-generativeai
nltk
tqdm
fastapi
uvicorn
python-multipart
python-dotenv
jinja2
pandas
scikit-learn
ipykernel
numpy
transformers